+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Cinecred Changelog
 - When different fonts are used within the same grid or flow row, the text may
   now be vertically justified in an arbitrary way, while previously it was just
   being baseline-aligned.
+- Added support for variable fonts.
 - Embedded videos can now loop.
 - The preview resolution of embedded videos is now configurable.
 - The credits runtime can now be copied to the clipboard by clicking on it.